complete registration, login, and logout functionality

+++ base/views.py
@@ -1,7 +1,61 @@
 from django.shortcuts import render, redirect 
+from django.contrib import messages
 from .models import Room, Topic 
 from .forms import RoomForm
 from django.db.models import Q 
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout 
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
+
+
+def loginPage(request):
+   
+    page = 'login'
+
+    if request.user.is_authenticated: 
+        return redirect('home')
+
+    if request.method == 'POST': 
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+        try: 
+            user = User.objects.get(username = username)
+        except: 
+            messages.error(request, "User does not exist. Please register")
+        user = authenticate(request, username=username, password=password)
+        if user is not None: 
+            login(request, user)
+            return redirect('home')
+        else: 
+            messages.error(request, 'Incorrect username or password. Sorry!')
+
+    context = {'page': page}
+    return render(request,'base/login_register.html', context)
+
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request): 
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else: 
+            messages.error(request, 'An error occured during registration, please try again')
+    context = {'form': form}
+    return render(request, 'base/login_register.html', context )
+
+
 def home(request):
     q = request.GET.get('q')
     if(q != None): q 
@@ -16,11 +70,16 @@ def home(request):
     context = {'rooms': rooms , 'topics': topics, 'room_count' : room_count }
     return render(request, 'base/home.html', context )
 
+
+
 def room(request, pk):
     room = Room.objects.get(id = pk)
     context = {'room' : room} 
     return render(request, 'base/room.html', context)
 
+
+
+@login_required(login_url="login")
 def createRoom(request):
     form = RoomForm()
     if(request.method == 'POST'): 
@@ -34,9 +93,14 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+
+@login_required(login_url="login")
 def updateRoom(request, pk): 
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.user:
+        return HttpResponse('You must be a room owner to make updates')
     
     if(request.method == 'POST'): 
         form = RoomForm(request.POST, instance=room)
@@ -47,8 +111,13 @@ def updateRoom(request, pk):
     context = { 'form': form} 
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.user:
+        return HttpResponse('You must be a room owner to delete rooms')
+    
     if(request.method == 'POST'): 
         room.delete() 
         return redirect('home')
